Allow operating on multiple repated fields when one of them only has 1 or 0 values
Summary:
In D33210, I made it so that we could apply functions to multiple repeated
fields so long as the repetition level matches as a proxy for allowing multiple
selections derived from the same repeated field.  However, there's an additional
case that bigquery allows: when one of the repeated fields actually only
contains values derived originally from scalar data or the same repeated field
again.  In these cases, the number of repeated elements doesn't match.

This appears to occur in cases like:
```
something_repeated + IF(some_condition, something_scalar, something_repeated)
```

However, as in D33210, we're not set up to track the provenance of individual
data points, so we can't actually match bigquery's behavior exactly.  Instead,
I've relaxed the requirement further to allow operations when there's 1 or 0
values in a repeated field too.  Again, this is more permissive than bigquery,
so I've added a TODO to fix this.

Test Plan: `python -m unittest evaluator_test`

Reviewers: ragini

Reviewed By: ragini

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D33261