Handle repeated fields better in WHERE clauses
Summary:
In a number of prior diffs, I changed handling of function calls on repeated
fields.  This revision implements a similar thing for WHERE clauses.

When evaluating a WHERE clause on a repeated field, we need to keep scalar
column rows if any of the repeated values is true, and for other repeated
fields, we match up the values within a row (as when applying a function on two
repeated columns).  As for repeated functions, this should only really be
allowed when the data is derived from the same field or from within the same
repeated record, but we can't determine this right now, so I've allowed it
generally and added a TODO to fix.  (Likewise, we need to accept repeated fields
with 0 or 1 value generally for the same reason as in repeated function
evaluation.)

In order to implement this, I refactored out some of the logic used for
evaluating functions on repeated fields into repeated_util.py and then reused it
for the masking.

Test Plan: `python -m unittest evaluator_test`

Reviewers: ragini

Reviewed By: ragini

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D33321