Add a simple setup.py file to tinyquery to allow local installation with pip
Summary:

In order to be able to use tinyquery as a package and not a submodule, this
commit adds a simple setup.py file.  In order to install systemwide, the steps
are:
1. From the tinyquery base directory `ln -s . tinyquery`
2. `python setup.py install`

Step (1) is needed because setup.py expects the package to be in its own
directory, and we have tinyquery set up so that the package is the top level
directory.  I'd like to change that eventually, but for now I'm trying to avoid
the refactor that would be necessary in webapp to resolve the path correctly
after changing that.

Test Plan:
- `ln -s . tinyquery`
- `python setup.py install`
- `python`
- `import tinyquery`
- verify there's no import error

Auditors: nabil