Move tinyquery code into a subdirectory
Summary:

Python's packaging tools expect the code for the tinyquery package to live in a
subdirectory called `tinyquery` of the root directory of the project.  This
commit moves the code and updates the lint blacklist accordingly.  It also
updates the .gitignore to ignore build artifacts.

This removes the need for the linking step of installation described in
948e7eab1bd0dcf01208e5a70aaa68ee45608765. (It turned out fixing the resolution
of this dependency in webapp wasn't hard at all.)

Test Plan:
`python -m unittest discover tinyquery --pattern=*_test.py`

Auditors: nabil