Python 3 iterators (#12)
Various changes to the code to support python3 (and python2):

* iteritems and itervalues replaced with items and values
* Python 3 zip support
* six for xrange

The first two have the potential to have perf issues, since for large items itervalues() can be much more efficient than values().  However, an audit of all the use-cases showed that we only use iteritems() and itervalues() for small things, typically to iterate over columns in a table.

zip returns a list in python2 but an iterator in python3.  This change casts to a list when necessary for compatibility. That's not too many places and again the items we are zipping are small.

xrange is a different story: there we *do* often iterate over things that can be quite large (like the number of rows) and do not want to use `range` in python2.  So it's time to bring in `six`, so we can use an iterator in each language.